django/gecommerce/requirements.txt
Django==2.1.7
djangorestframework
Pillow
psycopg2